Add legacy fields to ElasticSearch index
As part of debugging the issues between Query Parser and ElasticSearch,
we’ve decided to add some legacy fields and data structures that Query
Parser seems to expect when composing its query. This commit:

- Creates a utility function for decorating the existing index entries
  with legacy fields
- Creates a utilities file to contain this function
- Imports and uses this function to decorate the index before adding to
  ElasticSearch

Note that I’m unable to actually test this code since we have to
package and deploy it to AWS, so… care should be taken in reviewing.

Affects #1

Signed-off-by: Justin Reese <justin@justinreese.com>